fork block number format

+++ core/vm/evm.go
@@ -1373,7 +1373,7 @@ func (evm *EVM) HandleNFT(
 		Erb100000 := big.NewInt(70000)
 		Erb100000.Mul(Erb100000, baseErb)
 		empty := common.Address{}
-		if evm.Context.BlockNumber.Uint64() < uint64(types.SwitchBranchBlock) {
+		if evm.Context.BlockNumber.Uint64() < types.SwitchBranchBlock {
 			if evm.Context.VerifyPledgedBalance(evm.StateDB, addr, Erb100000) {
 				err := evm.Context.MinerBecome(evm.StateDB, addr, empty)
 				if err != nil {
@@ -1434,7 +1434,7 @@ func (evm *EVM) HandleNFT(
 	case 10: // cancel pledge of token
 		log.Info("HandleNFT(), CancelPledgedToken>>>>>>>>>>", "wormholes.Type", wormholes.Type,
 			"blocknumber", evm.Context.BlockNumber.Uint64())
-		if evm.Context.BlockNumber.Uint64() < uint64(types.SwitchBranchBlock) {
+		if evm.Context.BlockNumber.Uint64() < types.SwitchBranchBlock {
 			stakerpledged := evm.Context.GetStakerPledged(evm.StateDB, caller.Address(), addr)
 
 			baseErb, _ := new(big.Int).SetString("1000000000000000000", 10)
